Add all WEB!!!

+++ dao-api/src/main/java/com/itacademy/jd2/vvm/parking/dao/api/entity/table/ITariff.java
@@ -1,14 +1,16 @@
 package com.itacademy.jd2.vvm.parking.dao.api.entity.table;
 
+import java.math.BigDecimal;
+
 public interface ITariff extends IBaseEntity {
 
 	String getName();
 
 	void setName(String name);
 
-	Integer getPrice();
+	BigDecimal getPrice();
 
-	void setPrice(Integer price);
+	void setPrice(BigDecimal price);
 
 	String getStatus();
